Create day.c
This program was inspired by a prompt on Reddit, specifically from r/Programming Prompts.  
The program will calculate the day of the week of any inputted date, within bounds from March 1, 1900 to December 31, 2099.  
The date should be input as a command line argument, in the format MM DD YYYY, following the command ./day immediately.
The program will determine the day of the week of the given date and print it to the screen, then exit.